requirements.txt
langchain-core>=0.3
langchain-community>=0.2
langchain-groq>=0.2
python-dotenv>=1.0
pydantic>=2.6